DBZ-9453 Extract source/destination connection config from the Connection
